Refactor: inspectedItem property.
Added an inspectedItem in the TreeModel. The rootItem is always invisible. If the obj_name is the empty string, the inspectedItem will be the rootItem (and therefore be invisible). If the obj_name is given, an sible root item will be added and the inspectedItem will be its only child. In that case the inspected item will be visible.